move item editor to separate window
